requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
PyMySQL==1.1.1
cryptography==42.0.8
Werkzeug==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
click==8.1.7
itsdangerous==2.2.0
SQLAlchemy==2.0.32
requests==2.31.0 